+++ Utilities/tiff/tif_unix.c
@@ -35,13 +35,13 @@
 static tsize_t
 _tiffReadProc(thandle_t fd, tdata_t buf, tsize_t size)
 {
-        return ((tsize_t) read((int) fd, buf, (size_t) size));
+        return ((tsize_t) read(((int) (long int)fd), buf, (size_t) size));
 }
 
 static tsize_t
 _tiffWriteProc(thandle_t fd, tdata_t buf, tsize_t size)
 {
-        return ((tsize_t) write((int) fd, buf, (size_t) size));
+        return ((tsize_t) write(((int) (long int)fd), buf, (size_t) size));
 }
 
 static toff_t
@@ -50,14 +50,14 @@ _tiffSeekProc(thandle_t fd, toff_t off, int whence)
 #if USE_64BIT_API == 1
         return ((toff_t) lseek64((int) fd, (off64_t) off, whence));
 #else
-        return ((toff_t) lseek((int) fd, (off_t) off, whence));
+        return ((toff_t) lseek((int) (long int)fd, (off_t) off, whence));
 #endif
 }
 
 static int
 _tiffCloseProc(thandle_t fd)
 {
-        return (close((int) fd));
+        return (close((int) (long int)fd));
 }
 
 #include <sys/stat.h>
@@ -74,7 +74,7 @@ _tiffSizeProc(thandle_t fd)
         return (toff_t) (fstat64((int) fd, &sb) < 0 ? 0 : sb.st_size);
 #else
         struct stat sb;
-        return (toff_t) (fstat((int) fd, &sb) < 0 ? 0 : sb.st_size);
+        return (toff_t) (fstat((int) (long int)fd, &sb) < 0 ? 0 : sb.st_size);
 #endif
 #endif
 }
@@ -127,7 +127,7 @@ TIFFFdOpen(int fd, const char* name, const char* mode)
         TIFF* tif;
 
         tif = TIFFClientOpen(name, mode,
-            (thandle_t) fd,
+            (thandle_t) (long int)fd,
             _tiffReadProc, _tiffWriteProc,
             _tiffSeekProc, _tiffCloseProc, _tiffSizeProc,
             _tiffMapProc, _tiffUnmapProc);
